+++ google_analytics/utils.py
@@ -1,17 +1,20 @@
 import random
-import sys
 import time
 import uuid
 import hashlib
+import structlog
 
 from django.conf import settings
 from django.utils.translation import get_language_from_request
-from six.moves import urllib
 
 from google_analytics import CAMPAIGN_TRACKING_PARAMS
 
-if sys.version_info[0] == 3:
-    unicode = str
+from six import text_type
+from six.moves.urllib.parse import quote, urlencode
+try:
+    from urllib.parse import urlparse
+except ImportError:
+    from urlparse import urlparse
 
 VERSION = '1'
 COOKIE_NAME = '__utmmobile'
@@ -61,13 +64,21 @@ def set_cookie(params, response):
 
 
 def build_ga_params(
-        request, account, path=None, event=None, referer=None, title=None):
+        request, account, path=None, event=None, referer=None, title=None,
+        user_id=None, custom_params={}):
     meta = request.META
     # determine the domian
     domain = meta.get('HTTP_HOST', '')
+    if request.method == 'GET':
+        ni = '0'
+    else:
+        ni = '1'
 
     # determine the referrer
     referer = referer or request.GET.get('r', '')
+    parse_referer = urlparse(referer)
+    if parse_referer.netloc == domain:
+        referer = parse_referer.path
 
     custom_uip = None
     if hasattr(settings, 'CUSTOM_UIP_HEADER') and settings.CUSTOM_UIP_HEADER:
@@ -76,8 +87,14 @@ def build_ga_params(
     path = path or request.GET.get('p', '/')
 
     # get client ip address
-    if 'HTTP_X_FORWARDED_FOR' in meta:
+    if 'HTTP_X_FORWARDED_FOR' in meta and meta.get('HTTP_X_FORWARDED_FOR', ''):
         client_ip = meta.get('HTTP_X_FORWARDED_FOR', '')
+        if client_ip:
+            # The values in a proxied environment are usually presented in the
+            # following format:
+            # X-Forwarded-For: client, proxy1, proxy2
+            # In this case, we want the client IP Only
+            client_ip = client_ip.split(',')[0]
     else:
         client_ip = meta.get('REMOTE_ADDR', '')
 
@@ -95,17 +112,24 @@ def build_ga_params(
         'dh': domain,
         'sr': '',
         'dr': referer,
-        'dp': urllib.parse.quote(path.encode('utf-8')),
+        'dp': quote(path.encode('utf-8')),
         'tid': account,
         'cid': visitor_id,
         'uip': custom_uip or client_ip,
+        'ni': ni,
     }
+
+    # add user ID if exists
+    if user_id:
+        params.update({'uid': user_id})
+
+    # add custom parameters
+    params.update(custom_params)
 
     # add page title if supplied
     if title:
         u_title = title.decode('utf-8') if isinstance(title, bytes) else title
-        params.update(
-            {'dt': urllib.parse.quote(unicode(u_title).encode('utf-8'))})
+        params.update({'dt': quote(text_type(u_title).encode('utf-8'))})
     # add event parameters if supplied
     if event:
         params.update({
@@ -120,8 +144,9 @@ def build_ga_params(
 
     # update campaign params from request
     for param in CAMPAIGN_TRACKING_PARAMS:
-        if param in request.GET:
-            campaign_params[param] = request.GET[param]
+        ga_name = CAMPAIGN_TRACKING_PARAMS.get(param)
+        if ga_name in request.GET:
+            campaign_params[param] = request.GET[ga_name]
 
     # store campaign tracking parameters in session
     request.session[CAMPAIGN_PARAMS_KEY] = campaign_params
@@ -131,7 +156,14 @@ def build_ga_params(
 
     # construct the gif hit url
     ga_url = "http://www.google-analytics.com/collect"
-    utm_url = ga_url + "?&" + urllib.parse.urlencode(params)
+    utm_url = ga_url + "?&" + urlencode(params)
+    ga_logging_enabled = False
+    if hasattr(settings, 'ENABLE_GA_LOGGING'):
+        if settings.ENABLE_GA_LOGGING:
+            log = structlog.get_logger()
+            log.msg('GA_URL: %s' % utm_url, user_agent=user_agent)
+            ga_logging_enabled = True
+
     locale = get_language_from_request(request)
 
     return {'utm_url': utm_url,
@@ -142,4 +174,5 @@ def build_ga_params(
             'COOKIE_USER_PERSISTENCE': COOKIE_USER_PERSISTENCE,
             'COOKIE_NAME': COOKIE_NAME,
             'COOKIE_PATH': COOKIE_PATH,
+            'ga_logging_enabled': ga_logging_enabled
             }